backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
boto3>=1.35.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
alembic>=1.13.0
redis>=5.0.0
celery>=5.4.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pillow>=11.0.0
opencv-python-headless>=4.10.0
pandas>=2.2.0
pyarrow>=17.0.0
spacy>=3.8.0
nltk>=3.9.0
aiofiles>=24.1.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
